+++ rtw_hooks/t32xil_make_rtw_before_make_user_hook.m
@@ -1,11 +1,11 @@
 % --------------------------------------------------------------------------------
 % @Title: Customize build process of T32XIL
-% @Description: 
-%   Set custom user hooks that are activated during the build process of 
+% @Description:
+%   Set custom user hooks that are activated during the build process of
 %   T32XIL. Please refer to the section
 %   "Customize Build Process withsl_customization.m" for a detailed
 %   description of the configuration options.
-%   
+%
 % @Keywords: PIL Processor-in-the-Loop Simulink MATLAB model-based
 % @Author: CSA
 % @Copyright: (C) 1989-2017 Lauterbach GmbH, licensed for use with TRACE32(R) only
@@ -16,19 +16,21 @@ function t32xil_make_rtw_before_make_user_hook(modelName, dependencyObject)
 
 toolchain = get_param(modelName, 'Toolchain');
 if strcmp(toolchain, 'TRACE32 XIL GCC arm-none-eabi Cortex-R | gmake makefile')
-    base = fullfile(fileparts(mfilename('fullpath')), '..', 'targets', 'cortex-r');    
+    base = fullfile(fileparts(mfilename('fullpath')), '..', 'targets', 'cortex-r');
     addSourcePaths(dependencyObject, fullfile(base));
-    
+
     addSourceFiles(dependencyObject, 'crt0.s');
 
     addLinkFlags(dependencyObject, ['-T ' fullfile(base, 'link.lds')]);
 elseif strcmp(toolchain, 'TRACE32 XIL GCC arm-none-eabi Cortex-M | gmake makefile')
-    base = fullfile(fileparts(mfilename('fullpath')), '..', 'targets', 'cortex-m');    
+    base = fullfile(fileparts(mfilename('fullpath')), '..', 'targets', 'cortex-m');
     addSourcePaths(dependencyObject, fullfile(base));
-    
+
     addSourceFiles(dependencyObject, 'crt0.s');
 
-    addLinkFlags(dependencyObject, ['-T ' fullfile(base, 'link.lds')]);    
+    addLinkFlags(dependencyObject, ['-T ' fullfile(base, 'link.lds')]);
+elseif strcmp(toolchain, 'TRACE32 XIL Renesas RL78 Compiler CC-RL | gmake makefile')
+    base = fullfile('C:', 'Program Files (x86)', 'Renesas Electronics', 'CS+', 'CC', 'SampleProjects', 'RL78', 'RL78_G13_Tutorial_Basic_Operation_CC');
 end
 
 end