+++ toolchain/rtwTargetInfo.m
@@ -12,9 +12,9 @@
 %   
 % @Keywords: PIL Processor-in-the-Loop Simulink MATLAB model-based
 % @Author: CSA
-% @Copyright: (C) 1989-2019 Lauterbach GmbH, licensed for use with TRACE32(R) only
+% @Copyright: (C) 1989-2022 Lauterbach GmbH, licensed for use with TRACE32(R) only
 % --------------------------------------------------------------------------------
-% $Id: rtwTargetInfo.m 5825 2021-11-19 08:35:56Z csax $
+% $Id: rtwTargetInfo.m 6292 2022-06-02 14:25:35Z csax $
 
 %% Callback to detect the custom toolchains
 function rtwTargetInfo(tr)
@@ -32,13 +32,21 @@ function config = loc_createToolchain
     config(2) = coder.make.ToolchainInfoRegistry;
     config(2).Name = 'TRACE32 XIL GCC arm-none-eabi Cortex-R | gmake makefile';
     config(2).FileName = fullfile(fileparts(mfilename('fullpath')), 't32xil_tc_gcc_arm_none_eabi_cortex_r.mat');
-    config(2).TargetHWDeviceType = {'ARM Compatible->ARM Cortex'};
+    if verLessThan('matlab', '9.12')
+        config(2).TargetHWDeviceType = {'ARM Compatible->ARM Cortex'};
+    else
+        config(2).TargetHWDeviceType = {'ARM Compatible->ARM Cortex-R'};
+    end
     config(2).Platform  = {computer('arch')};
 
     config(3) = coder.make.ToolchainInfoRegistry;
     config(3).Name = 'TRACE32 XIL GCC arm-none-eabi Cortex-M | gmake makefile';
     config(3).FileName = fullfile(fileparts(mfilename('fullpath')), 't32xil_tc_gcc_arm_none_eabi_cortex_m.mat');
-    config(3).TargetHWDeviceType = {'ARM Compatible->ARM Cortex'};
+    if verLessThan('matlab', '9.12')
+        config(3).TargetHWDeviceType = {'ARM Compatible->ARM Cortex'};
+    else
+        config(3).TargetHWDeviceType = {'ARM Compatible->ARM Cortex-M'};
+    end
     config(3).Platform  = {computer('arch')};
 
     config(4) = coder.make.ToolchainInfoRegistry;
@@ -52,6 +60,12 @@ function config = loc_createToolchain
     config(5).FileName = fullfile(fileparts(mfilename('fullpath')), 't32xil_tc_hightec_tricore_gcc.mat');
     config(5).TargetHWDeviceType = {'Infineon->TriCore'};
     config(5).Platform  = {computer('arch')};
+
+    config(6) = coder.make.ToolchainInfoRegistry;
+    config(6).Name = 'TRACE32 XIL Renesas RL78 Compiler CC-RL | gmake makefile';
+    config(6).FileName = fullfile(fileparts(mfilename('fullpath')), 't32xil_tc_renesas_rl78_ccrl.mat');
+    config(6).TargetHWDeviceType = {'Renesas->RL78'};
+    config(6).Platform  = {computer('arch')};
 
     % To register more custom toolchains:
     % 1) Copy and paste the five preceding 'config' lines.